Add accessToken(가게 저장,삭제,수정,내 가게 가져오기 시 필요하다),

+++ api/throwapi/src/main/java/sosteam/throwapi/domain/store/service/StoreDeleteService.java
@@ -5,10 +5,13 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import sosteam.throwapi.domain.store.entity.Address;
 import sosteam.throwapi.domain.store.entity.Store;
-import sosteam.throwapi.domain.store.entity.dto.StoreDto;
-import sosteam.throwapi.domain.store.exception.InvalidDeleteException;
+import sosteam.throwapi.domain.store.exception.InvalidRequestException;
 import sosteam.throwapi.domain.store.exception.NoSuchStoreException;
 import sosteam.throwapi.domain.store.repository.repo.StoreRepository;
+import sosteam.throwapi.domain.user.entity.User;
+import sosteam.throwapi.domain.user.entity.dto.user.UserInfoDto;
+import sosteam.throwapi.domain.user.service.UserInfoService;
+import sosteam.throwapi.global.service.JwtTokenService;
 
 import java.util.Optional;
 import java.util.UUID;
@@ -18,22 +21,36 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class StoreDeleteService {
     private final StoreRepository storeRepository;
+    private final UserInfoService userInfoService;
+    private final JwtTokenService jwtTokenService;
 
 
-    public void deleteStore(UUID extStoreId) {
+    public void deleteStore(String accessToken, UUID extStoreId) {
+        // 요청 사용자 정보 가져오기
+        UserInfoDto userInfoDto = new UserInfoDto(
+                jwtTokenService.extractSubject(accessToken)
+        );
+        User user = userInfoService.searchByInputId(userInfoDto);
+
         Optional<Store> optionalStore = storeRepository.searchByExtStoreId(extStoreId);
         if(optionalStore.isEmpty()) throw new NoSuchStoreException();
         Store store = optionalStore.get();
+
+        // check if the store's owner is same with request-user
+        Optional<UUID> userId = storeRepository.searchUserByStore(store);
+        if(!user.getId().equals(userId)) throw new InvalidRequestException("DELETE");
+
+
         Optional<Address> optionalAddress = storeRepository.searchAddressByStore(store.getId());
         if(optionalAddress.isEmpty()) throw new NoSuchStoreException();
         Address address = optionalAddress.get();
-        // 요청이 들어온 데이터와 삭제하고자하는 데이터가 전부 일치해야한다.
+
         if(
            store.getExtStoreId().equals(extStoreId)
         ) {
             store.modifyAddress(null);
             address.modifyStore(null);
             storeRepository.delete(store);
-        } else throw new InvalidDeleteException();
+        } else throw new InvalidRequestException("DELETE");
     }
 }
